Исправляет баги и добавляет оптимизацию кода
